Moved all *ViewModel classes to viewmodel package; Renamed ProfileViewModel to LogOutViewModel, LoginViewModel to LogInViewModel, SplashViewModel to CheckExistingSessionViewModel classes
